refactor: made functions handle biocyc pattern
The biocyc ID pattern for both metablites and reactions match
other IDs aswell. However, this seems to be only the case, because
the biocyc pattern is fairly broad. It doesn't seem to be true for
any other pattern, hence I implemented a clean-up step in the
dataframes output by collect_{rxn|met}_id_namespace.