refactor: change match order of IDs
I ordered the dictionaries REACTION_ANNOTATIONS and METABOLITE_ANNOTATIONS so that the patterns will be applied in
a non-overlapping fashion. I also altered the collect functions
to break as soon as a pattern matches an ID. This is is to avoid
duplicate matching, since BiGG and BioCyc patterns are very
generic and are very likely to match all the other identifiers too.