feat: add import function gene essentiality data
The added function imports all .csv files in a specified
directory that are prefixed with 'e_'. The data from all the files
is then collated into a dataframe, in which the data from individual files is stored as individual columns with the gene IDs as the index.
We assume that researchers create a file per essentiality experiment
based on different conditions.