Fix/add metanetx id check (#392)
* feat: add MNX id for O2

This script auto-generates MNX id json files, so needs to be updated with MNX id for O2 and run to auto-generate updated MNX id shortlist in the reference json file

* feat: add O2 to met_id_shortlist.json

This was done using generate_mnx_shortlist.py

* fix: fix function to use MNX id

find_oxygen_reactions now checks MetaNetX id in addition to checking metabolite formulas for O2

* fix: function to return rxns

find_oxygen_reactions returned reactions and metabolites, it has now been fixed to only return reactions

* refactor: reduce list comprehension from 2 to 1

* fix: add MNXM20 to shortlist

MNXM20 is an essential component needed in the shortlist for detecting energy-coupling tests. This fix adds it back in

* fix: fix json file

This updates the json file based on the changes to generate_mnx_shortlist.py in the last commit. Namely mnx20 is now included in the json file

* chore: add commit changes to HISTORY.rst